api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
celery>=5.3.0
redis>=5.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
# Authentication
authlib==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
# HTTP client for OAuth
httpx==0.26.0
# Email service
sendgrid==6.11.0

